requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.5.1
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
Django==3.1.7
djangorestframework==3.12.2
idna==2.10
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2==2.8.6
pylint==2.7.2
pytz==2021.1
regex==2020.11.13
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
wrapt==1.12.1
